+++ trunk/src/utest/srs_utest_config.cpp
@@ -2675,7 +2675,7 @@ VOID TEST(ConfigMainTest, CheckGlobalConfig)
     if (true) {
         MockSrsConfig conf;
         HELPER_ASSERT_SUCCESS(conf.parse("listen 1935;"));
-        EXPECT_EQ(1, conf.get_listens().size());
+        EXPECT_EQ(1, (int)conf.get_listens().size());
     }
 
     if (true) {
